requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/dev.txt requirements.in/dev.in
#
anyascii==0.1.7
    # via
    #   -r requirements.in/../requirements/base.txt
    #   wagtail
appdirs==1.4.4
    # via black
attrs==20.2.0
    # via black
autopep8==1.4.4
    # via -r requirements.in/../requirements/base.txt
beautifulsoup4==4.8.2
    # via
    #   -r requirements.in/../requirements/base.txt
    #   wagtail
black==19.10b0
    # via -r requirements.in/dev.in
boto3==1.9.197
    # via -r requirements.in/../requirements/base.txt
botocore==1.12.197
    # via
    #   -r requirements.in/../requirements/base.txt
    #   boto3
    #   s3transfer
certifi==2019.6.16
    # via
    #   -r requirements.in/../requirements/base.txt
    #   requests
chardet==3.0.4
    # via
    #   -r requirements.in/../requirements/base.txt
    #   requests
click==7.1.2
    # via
    #   black
    #   pip-tools
coverage==5.2
    # via -r requirements.in/dev.in
dj-database-url==0.5.0
    # via -r requirements.in/../requirements/base.txt
django-environ==0.4.5
    # via -r requirements.in/../requirements/base.txt
django-filter==2.4.0
    # via
    #   -r requirements.in/../requirements/base.txt
    #   wagtail
django-modelcluster==5.1
    # via
    #   -r requirements.in/../requirements/base.txt
    #   wagtail
django-oauth-toolkit==1.2.0
    # via -r requirements.in/../requirements/base.txt
django-staff-sso-client==1.0.0
    # via -r requirements.in/../requirements/base.txt
django-storages==1.7.1
    # via -r requirements.in/../requirements/base.txt
django-taggit==1.2.0
    # via
    #   -r requirements.in/../requirements/base.txt
    #   wagtail
django-treebeard==4.3
    # via
    #   -r requirements.in/../requirements/base.txt
    #   wagtail
django==2.2.20
    # via
    #   -r requirements.in/../requirements/base.txt
    #   django-filter
    #   django-oauth-toolkit
    #   django-staff-sso-client
    #   django-storages
    #   django-taggit
    #   django-treebeard
    #   djangorestframework
    #   wagtail
djangorestframework==3.12.4
    # via
    #   -r requirements.in/../requirements/base.txt
    #   wagtail
docutils==0.14
    # via
    #   -r requirements.in/../requirements/base.txt
    #   botocore
draftjs-exporter==2.1.6
    # via
    #   -r requirements.in/../requirements/base.txt
    #   wagtail
et-xmlfile==1.0.1
    # via
    #   -r requirements.in/../requirements/base.txt
    #   openpyxl
flake8==3.8.3
    # via -r requirements.in/dev.in
html5lib==1.0.1
    # via
    #   -r requirements.in/../requirements/base.txt
    #   wagtail
idna==2.8
    # via
    #   -r requirements.in/../requirements/base.txt
    #   requests
importlib-metadata==3.10.0
    # via
    #   flake8
    #   pep517
jmespath==0.9.4
    # via
    #   -r requirements.in/../requirements/base.txt
    #   boto3
    #   botocore
l18n==2018.5
    # via
    #   -r requirements.in/../requirements/base.txt
    #   wagtail
mccabe==0.6.1
    # via flake8
oauthlib==3.0.2
    # via
    #   -r requirements.in/../requirements/base.txt
    #   django-oauth-toolkit
    #   requests-oauthlib
openpyxl==3.0.7
    # via
    #   -r requirements.in/../requirements/base.txt
    #   tablib
pathspec==0.8.0
    # via black
pep517==0.10.0
    # via pip-tools
pillow==8.2.0
    # via
    #   -r requirements.in/../requirements/base.txt
    #   wagtail
pip-tools==6.0.1
    # via -r requirements.in/dev.in
psycopg2==2.8.3
    # via -r requirements.in/../requirements/base.txt
pycodestyle==2.6.0
    # via
    #   -r requirements.in/../requirements/base.txt
    #   autopep8
    #   flake8
pyflakes==2.2.0
    # via flake8
python-dateutil==2.8.0
    # via
    #   -r requirements.in/../requirements/base.txt
    #   botocore
pytz==2019.2
    # via
    #   -r requirements.in/../requirements/base.txt
    #   django
    #   django-modelcluster
    #   l18n
raven==6.10.0
    # via
    #   -r requirements.in/../requirements/base.txt
    #   django-staff-sso-client
regex==2020.10.23
    # via black
requests-oauthlib==1.2.0
    # via
    #   -r requirements.in/../requirements/base.txt
    #   django-staff-sso-client
requests==2.22.0
    # via
    #   -r requirements.in/../requirements/base.txt
    #   django-oauth-toolkit
    #   requests-oauthlib
    #   wagtail
s3transfer==0.2.1
    # via
    #   -r requirements.in/../requirements/base.txt
    #   boto3
six==1.12.0
    # via
    #   -r requirements.in/../requirements/base.txt
    #   html5lib
    #   l18n
    #   python-dateutil
soupsieve==2.0.1
    # via
    #   -r requirements.in/../requirements/base.txt
    #   beautifulsoup4
sqlparse==0.3.0
    # via
    #   -r requirements.in/../requirements/base.txt
    #   django
tablib[xls,xlsx]==3.0.0
    # via
    #   -r requirements.in/../requirements/base.txt
    #   wagtail
toml==0.10.1
    # via
    #   black
    #   pep517
typed-ast==1.4.1
    # via black
typing-extensions==3.7.4.3
    # via importlib-metadata
unidecode==1.1.1
    # via -r requirements.in/../requirements/base.txt
urllib3==1.25.3
    # via
    #   -r requirements.in/../requirements/base.txt
    #   botocore
    #   requests
wagtail==2.12.3
    # via
    #   -r requirements.in/../requirements/base.txt
    #   wagtailcodeblock
wagtailcodeblock==1.17.1.0
    # via -r requirements.in/../requirements/base.txt
waitress==1.4.3
    # via -r requirements.in/../requirements/base.txt
wand==0.5.5
    # via -r requirements.in/../requirements/base.txt
webencodings==0.5.1
    # via
    #   -r requirements.in/../requirements/base.txt
    #   html5lib
whitenoise==1.0.6
    # via -r requirements.in/../requirements/base.txt
willow==1.4
    # via
    #   -r requirements.in/../requirements/base.txt
    #   wagtail
xlrd==2.0.1
    # via
    #   -r requirements.in/../requirements/base.txt
    #   tablib
xlsxwriter==1.2.8
    # via
    #   -r requirements.in/../requirements/base.txt
    #   wagtail
xlwt==1.3.0
    # via
    #   -r requirements.in/../requirements/base.txt
    #   tablib
zipp==3.4.1
    # via
    #   importlib-metadata
    #   pep517

# The following packages are considered to be unsafe in a requirements file:
# pip
